updates moved all code for questPage to socials deleted older questPage
