+++ frontend/src/app/app/quests/socials/components/TasksMobile.tsx
