+++ frontend/src/app/app/quests/socials/page.tsx
@@ -1,0 +1,51 @@
+// app/QuestPage/page.tsx
+'use client'
+import { useContext, useState } from "react";
+// import { ChevronDown } from "lucide-react";
+
+import { DarkModeContext } from "./components/DarkMode";
+import Task from './components/Tasks'
+import TaskMobile from './components/TasksMobile'
+interface Quest {
+    id: number;
+    title: string;
+    description: string;
+    reward: string;
+    status: 'available' | 'in_progress' | 'completed';
+}
+
+export default function socials() {
+    const { isDarkMode } = useContext(DarkModeContext);
+
+
+    const [isOpen, setIsOpen] = useState<Boolean>(false);
+
+    const toggleOpen = () => {
+        setIsOpen(!isOpen);
+    };
+
+    return (
+        <section className="md:p-4 flex flex-col gap-6">
+            <div className='flex flex-col gap-2'>
+                <h1 className={`font-[600] xl:text-[3.1rem] lg:text-[2.9rem] md:text-[2.8rem] sm:text-[2rem] xs:text-[1.5rem]  ${isDarkMode ? "  text-white" : "   text-black"}`}>Tasks</h1>
+
+                <p className={` xl:max-w-[70%] opacity-60 md:w-[80%] md:text-lg text-[1rem] ${isDarkMode ? "  text-white" : "   text-black"}`}>Get rewarded for completing Quests</p>
+
+            </div>
+
+            <div className="hidden lg:grid gap-6 md:grid-cols-1 lg:grid-rows-4">
+
+                 {/* this is a component for the task made it reuseable  */}
+                 <Task title='Partner Tasks' completedText="Completed 1/4 tasks"/>
+            </div>
+                <div className='flex flex-col lg:hidden gap-4'>
+            {/* this is a component for the task  made it reuseable for mobile  */}
+                 <TaskMobile title='Follow @peer_protocol on 𝕏 ' task='perform'/>
+                 <TaskMobile title='Follow @peer_protocol on 𝕏 ' task='pending'/>
+                 <TaskMobile title='Follow @peer_protocol on 𝕏 ' task='completed'/>
+                
+                </div>
+            
+        </section>
+    )
+}
