+++ frontend/src/app/app/quests/socials/quest.md
@@ -1,0 +1,1 @@
+## QUEST PAGE FOLDER STRUCTURE
