--- frontend/src/app/questPage/quest.md
@@ -1,9 +1,0 @@
-## QUEST PAGE FOLDER STRUCTURE
-
-QuestHome
-Questmain
-
-routes 
-socialQuest
-leaderBoard
-referrals
